requirements_new.txt
docling==1.0.0
cohere==5.11.1
pinecone==5.3.1
pinecone-client==0.5.4
pinecone-text==0.5.4
langchain==0.1.16 
langchain_cohere==0.1.2 
tiktoken==0.6.0
markdown-analysis==0.0.5
google
beautifulsoup4
jsonlines==4.0.0
python_dotenv==1.0.1
keybert==0.8.5
keyphrase-vectorizers==0.0.13
transformers==4.45.2
llama-index==0.11.18
llama-index-llms-cohere==0.3.1
llama-index-embeddings-cohere==0.2.1
llama-index-finetuning==0.2.1
datasets==3.0.1
llama-index-embeddings-huggingface==0.3.1
FlagEmbedding==1.2.11
peft==0.13.2
bitsandbytes==0.44.1
accelerate==1.0.1
trl==0.11.4
sentencepiece==0.2.0
wandb==0.18.5
flash-attn==2.6.3
huggingface_hub===0.24.7
ragas==0.2.1
langchain-cohere==0.3.1
